feat: make it easy to recreate a router
